backend/render_requirements.txt
fastapi
uvicorn[standard]
google-generativeai
motor
python-dotenv
pydantic
requests
httpx
GitPython
openai
